# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApproveProposalsRequest": {
        "additionalProperties": false,
        "description": "The candidates a person chose. An empty list means 'none of them', and is respected.",
        "properties": {
          "approved_names": {
            "items": {
              "type": "string"
            },
            "title": "Approved Names",
            "type": "array"
          }
        },
        "title": "ApproveProposalsRequest",
        "type": "object"
      },
...
      "CreateExperimentRequest": {
        "additionalProperties": false,
        "properties": {
          "dataset_uri": {
            "description": "S3 URI of an approved dataset (s3://bucket/key).",
            "title": "Dataset Uri",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "file_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override the inferred format.",
            "title": "File Format"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "problem_type": {
            "$ref": "#/components/schemas/RequestedProblemType",
            "default": "auto"
          },
          "target_column": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Target Column",
            "type": "string"
          }
        },
        "required": [
          "name",
          "dataset_uri",
          "target_column"
        ],
        "title": "CreateExperimentRequest",
        "type": "object"
      },
      "CreateExperimentResponse": {
        "additionalProperties": false,
        "properties": {
          "artifact_prefix": {
            "title": "Artifact Prefix",
            "type": "string"
          },
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ExperimentStatus"
          }
        },
        "required": [
          "experiment_id",
          "status",
          "artifact_prefix"
        ],
        "title": "CreateExperimentResponse",
        "type": "object"
      },
...
      "DateDifferenceProposal": {
        "additionalProperties": false,
        "description": "Elapsed time between two date columns of the dataset.\n\nBoth endpoints are columns. There is deliberately no \"difference from today\": that would\nproduce a different value on every run and make the experiment irreproducible.",
        "properties": {
          "available_at_prediction_time": {
            "description": "Whether every input is known when a prediction would actually be made. The proposer states its assumption; a human confirms it. No dictionary can settle this, and getting it wrong is how a good validation score fails in production.",
            "title": "Available At Prediction Time",
            "type": "boolean"
          },
          "end": {
            "title": "End",
            "type": "string"
          },
          "name": {
            "description": "Name of the derived column. Must not already exist.",
            "title": "Name",
            "type": "string"
          },
          "op": {
            "const": "date_difference",
            "default": "date_difference",
            "title": "Op",
            "type": "string"
          },
          "rationale": {
            "description": "Why this feature should help. Shown to the reviewer.",
            "title": "Rationale",
            "type": "string"
          },
          "start": {
            "title": "Start",
            "type": "string"
          },
          "unit": {
            "$ref": "#/components/schemas/DateUnit",
            "default": "days"
          }
        },
        "required": [
          "name",
          "rationale",
          "available_at_prediction_time",
          "start",
          "end"
        ],
        "title": "DateDifferenceProposal",
        "type": "object"
      },
      "DateUnit": {
        "enum": [
          "days",
          "months",
          "years"
        ],
        "title": "DateUnit",
        "type": "string"
      },
...
      "DifferenceProposal": {
        "additionalProperties": false,
        "description": "``left - right`` \u2014 gaps, headroom, unused allowance.",
        "properties": {
          "available_at_prediction_time": {
            "description": "Whether every input is known when a prediction would actually be made. The proposer states its assumption; a human confirms it. No dictionary can settle this, and getting it wrong is how a good validation score fails in production.",
            "title": "Available At Prediction Time",
            "type": "boolean"
          },
          "left": {
            "title": "Left",
            "type": "string"
          },
          "name": {
            "description": "Name of the derived column. Must not already exist.",
            "title": "Name",
            "type": "string"
          },
          "op": {
            "const": "difference",
            "default": "difference",
            "title": "Op",
            "type": "string"
          },
          "rationale": {
            "description": "Why this feature should help. Shown to the reviewer.",
            "title": "Rationale",
            "type": "string"
          },
          "right": {
            "title": "Right",
            "type": "string"
          }
        },
        "required": [
          "name",
          "rationale",
          "available_at_prediction_time",
          "left",
          "right"
        ],
        "title": "DifferenceProposal",
        "type": "object"
      },
...
      "ExperimentListResponse": {
        "additionalProperties": false,
        "properties": {
          "available_roots": {
            "description": "Artifact buckets this platform may read.",
            "items": {
              "type": "string"
            },
            "title": "Available Roots",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "experiments": {
            "items": {
              "$ref": "#/components/schemas/ExperimentRecord"
            },
            "title": "Experiments",
            "type": "array"
          },
          "root": {
            "default": "",
            "description": "The artifact bucket these experiments were read from.",
            "title": "Root",
            "type": "string"
          }
        },
        "required": [
          "experiments",
          "count"
        ],
        "title": "ExperimentListResponse",
        "type": "object"
      },
      "ExperimentRecord": {
        "additionalProperties": false,
        "description": "The composed view returned by the API. Never stored as one object.",
        "properties": {
          "artifact_prefix": {
            "title": "Artifact Prefix",
            "type": "string"
          },
          "best_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Model"
          },
          "best_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Score"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "current_stage": {
            "default": "created",
            "title": "Current Stage",
            "type": "string"
          },
          "dataset": {
            "$ref": "#/components/schemas/DatasetReference"
          },
          "deleted": {
            "default": false,
            "title": "Deleted",
            "type": "boolean"
          },
          "eda_execution_arn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eda Execution Arn"
          },
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "failure_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Message"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "primary_metric": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Metric"
          },
          "report_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Uri"
          },
          "requested_models": {
            "items": {
              "type": "string"
            },
            "title": "Requested Models",
            "type": "array"
          },
          "requested_problem_type": {
            "$ref": "#/components/schemas/RequestedProblemType",
            "default": "auto"
          },
          "status": {
            "$ref": "#/components/schemas/ExperimentStatus",
            "default": "CREATED"
          },
          "target_column": {
            "title": "Target Column",
            "type": "string"
          },
          "training_execution_arn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Execution Arn"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "experiment_id",
          "name",
          "created_at",
          "updated_at",
          "dataset",
          "target_column",
          "artifact_prefix"
        ],
        "title": "ExperimentRecord",
        "type": "object"
      },
      "ExperimentStatus": {
        "enum": [
          "CREATED",
          "EDA_RUNNING",
          "EDA_COMPLETED",
          "FEATURE_REVIEW",
          "READY_FOR_TRAINING",
          "PREPARING",
          "TRAINING",
          "EVALUATING",
          "COMPLETED",
          "COMPLETED_WITH_WARNINGS",
          "FAILED"
        ],
        "title": "ExperimentStatus",
        "type": "string"
      },
...
      "FeatureOp": {
        "enum": [
          "ratio",
          "difference",
          "date_difference",
          "map_categories",
          "is_missing"
        ],
        "title": "FeatureOp",
        "type": "string"
      },
      "FeatureProposalReport": {
        "additionalProperties": false,
        "description": "Everything proposed for one experiment and what happened to it.",
        "properties": {
          "approved_names": {
            "description": "The candidates a person approved. Validation decides what may be offered; this decides what is used. Empty until someone chooses.",
            "items": {
              "type": "string"
            },
            "title": "Approved Names",
            "type": "array"
          },
          "candidates": {
            "description": "Specifications that passed validation and are offered for review.",
            "items": {
              "discriminator": {
                "mapping": {
                  "date_difference": "#/components/schemas/DateDifferenceProposal",
                  "difference": "#/components/schemas/DifferenceProposal",
                  "is_missing": "#/components/schemas/IsMissingProposal",
                  "map_categories": "#/components/schemas/MapCategoriesProposal",
                  "ratio": "#/components/schemas/RatioProposal"
                },
                "propertyName": "op"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RatioProposal"
                },
                {
                  "$ref": "#/components/schemas/DifferenceProposal"
                },
                {
                  "$ref": "#/components/schemas/DateDifferenceProposal"
                },
                {
                  "$ref": "#/components/schemas/MapCategoriesProposal"
                },
                {
                  "$ref": "#/components/schemas/IsMissingProposal"
                }
              ]
            },
            "title": "Candidates",
            "type": "array"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decided_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By"
          },
          "experiment_id": {
            "title": "Experiment Id",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "proposed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Model id that produced the proposals; null when they were written by hand.",
            "title": "Proposed By"
          },
          "rejected": {
            "description": "Specifications refused before anyone saw them, and why.",
            "items": {
              "$ref": "#/components/schemas/RejectedProposal"
            },
            "title": "Rejected",
            "type": "array"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "experiment_id",
          "generated_at"
        ],
        "title": "FeatureProposalReport",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IsMissingProposal": {
        "additionalProperties": false,
        "description": "Flag whether a column is missing.\n\nMissingness that predicts the outcome is already a leakage finding; this turns the pattern\ninto an explicit feature, so the derived column is screened like any other.",
        "properties": {
          "available_at_prediction_time": {
            "description": "Whether every input is known when a prediction would actually be made. The proposer states its assumption; a human confirms it. No dictionary can settle this, and getting it wrong is how a good validation score fails in production.",
            "title": "Available At Prediction Time",
            "type": "boolean"
          },
          "column": {
            "title": "Column",
            "type": "string"
          },
          "name": {
            "description": "Name of the derived column. Must not already exist.",
            "title": "Name",
            "type": "string"
          },
          "op": {
            "const": "is_missing",
            "default": "is_missing",
            "title": "Op",
            "type": "string"
          },
          "rationale": {
            "description": "Why this feature should help. Shown to the reviewer.",
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "name",
          "rationale",
          "available_at_prediction_time",
          "column"
        ],
        "title": "IsMissingProposal",
        "type": "object"
      },
...
      "MapCategoriesProposal": {
        "additionalProperties": false,
        "description": "Group category values by documented meaning \u2014 country to region, code to family.\n\nThis is the operation a data dictionary actually unlocks: the mapping comes from what the\nvalues *mean*, which no statistic in the profile can supply.",
        "properties": {
          "available_at_prediction_time": {
            "description": "Whether every input is known when a prediction would actually be made. The proposer states its assumption; a human confirms it. No dictionary can settle this, and getting it wrong is how a good validation score fails in production.",
            "title": "Available At Prediction Time",
            "type": "boolean"
          },
          "column": {
            "title": "Column",
            "type": "string"
          },
          "default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Group for values absent from the mapping; null leaves them missing.",
            "title": "Default"
          },
          "mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Mapping",
            "type": "object"
          },
          "name": {
            "description": "Name of the derived column. Must not already exist.",
            "title": "Name",
            "type": "string"
          },
          "op": {
            "const": "map_categories",
            "default": "map_categories",
            "title": "Op",
            "type": "string"
          },
          "rationale": {
            "description": "Why this feature should help. Shown to the reviewer.",
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "name",
          "rationale",
          "available_at_prediction_time",
          "column",
          "mapping"
        ],
        "title": "MapCategoriesProposal",
        "type": "object"
      },
...
      "ProposeFeaturesRequest": {
        "additionalProperties": false,
        "description": "Business context the profile cannot supply. Both fields are optional but both help.",
        "properties": {
          "prediction_timing": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the prediction is made relative to the outcome. This is what lets the model judge whether a proposed feature would be known at that moment.",
            "title": "Prediction Timing"
          },
          "target_definition": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What the target actually means in business terms.",
            "title": "Target Definition"
          }
        },
        "title": "ProposeFeaturesRequest",
        "type": "object"
      },
...
      "RatioProposal": {
        "additionalProperties": false,
        "description": "``numerator / denominator`` \u2014 per-unit, per-month, per-capita quantities.",
        "properties": {
          "available_at_prediction_time": {
            "description": "Whether every input is known when a prediction would actually be made. The proposer states its assumption; a human confirms it. No dictionary can settle this, and getting it wrong is how a good validation score fails in production.",
            "title": "Available At Prediction Time",
            "type": "boolean"
          },
          "denominator": {
            "title": "Denominator",
            "type": "string"
          },
          "name": {
            "description": "Name of the derived column. Must not already exist.",
            "title": "Name",
            "type": "string"
          },
          "numerator": {
            "title": "Numerator",
            "type": "string"
          },
          "on_zero_denominator": {
            "$ref": "#/components/schemas/ZeroDenominatorPolicy",
            "default": "null"
          },
          "op": {
            "const": "ratio",
            "default": "ratio",
            "title": "Op",
            "type": "string"
          },
          "rationale": {
            "description": "Why this feature should help. Shown to the reviewer.",
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "name",
          "rationale",
          "available_at_prediction_time",
          "numerator",
          "denominator"
        ],
        "title": "RatioProposal",
        "type": "object"
      },
...
      "RejectedProposal": {
        "additionalProperties": false,
        "description": "A proposal that will not be computed, and why. Shown to the reviewer, never silent.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "op": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FeatureOp"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when the specification did not parse as any operation."
          },
          "reason": {
            "$ref": "#/components/schemas/ProposalRejection"
          }
        },
        "required": [
          "name",
          "reason",
          "message"
        ],
        "title": "RejectedProposal",
        "type": "object"
      },
      "RequestedProblemType": {
        "description": "What the user may ask for. ``auto`` is resolved during profiling.",
        "enum": [
          "auto",
          "binary_classification",
          "multiclass_classification",
          "regression"
        ],
        "title": "RequestedProblemType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ZeroDenominatorPolicy": {
        "description": "What a ratio does where the denominator is zero. Never silently infinite.",
        "enum": [
          "null",
          "zero"
        ],
        "title": "ZeroDenominatorPolicy",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "\nInternal platform that accelerates classical-ML POCs: automated EDA, deterministic data\nquality and leakage screening, supervised feature review, classical model training and\ncomparison, and \u2014 strictly after the deterministic results exist \u2014 semantic interpretation.\n",
    "title": "ML Factory",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/experiments": {
      "get": {
        "description": "Experiments found under an artifact bucket, newest first.",
        "operationId": "list_experiments_api_v1_experiments_get",
        "parameters": [
          {
            "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
            "in": "query",
            "name": "root",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Artifact bucket URI to read from. Defaults to this platform's bucket.",
              "examples": [
                "s3://my-ml-factory-artifacts"
              ],
              "title": "Root"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Restrict to experiments you created.",
            "in": "query",
            "name": "mine",
            "required": false,
            "schema": {
              "default": false,
              "description": "Restrict to experiments you created.",
              "title": "Mine",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-Remote-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Remote-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiments",
        "tags": [
          "experiments"
        ]
      },
      "post": {
        "description": "Create an experiment and start EDA. Returns immediately \u2014 AWS work runs asynchronously.",
        "operationId": "create_experiment_api_v1_experiments_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Remote-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Remote-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateExperimentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Experiment",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/api/v1/experiments/{experiment_id}/feature-proposals": {
      "get": {
        "operationId": "get_feature_proposals_api_v1_experiments__experiment_id__feature_proposals_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureProposalReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Feature Proposals",
        "tags": [
          "feature proposals"
        ]
      },
      "post": {
        "description": "Ask for derived-feature specifications. Nothing is used until someone approves it.",
        "operationId": "propose_features_api_v1_experiments__experiment_id__feature_proposals_post",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeFeaturesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureProposalReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose Features",
        "tags": [
          "feature proposals"
        ]
      },
      "put": {
        "description": "Record which proposals a person approved. Only these are frozen into the run.",
        "operationId": "approve_feature_proposals_api_v1_experiments__experiment_id__feature_proposals_put",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Remote-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Remote-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveProposalsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureProposalReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Feature Proposals",
        "tags": [
          "feature proposals"
        ]
      }
    },
...
  }
}